Extract SDL GL bootstrap from GlobalRendering.
Move window, context, and SDL video lifecycle code into its own translation unit so the render startup path has a cleaner seam without changing behavior.

Made-with: Cursor
