+++ rts/Rendering/CMakeLists.txt
@@ -67,6 +67,7 @@ set(sources_engine_Rendering
 		"${CMAKE_CURRENT_SOURCE_DIR}/GL/State.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/GL/myGL.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/GlobalRendering.cpp"
+		"${CMAKE_CURRENT_SOURCE_DIR}/Platform/SDLGLContext.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/GroundFlash.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/CommandDrawer.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/HUDDrawer.cpp"
