+++ rts/Rendering/GlobalRendering.cpp
@@ -395,253 +395,8 @@ void CGlobalRendering::PreKill()
 }
 
 
-SDL_Window* CGlobalRendering::CreateSDLWindow(const char* title) const
-{
-	SDL_Window* newWindow = nullptr;
-
-	const std::array aaLvls = {msaaLevel, msaaLevel / 2, msaaLevel / 4, msaaLevel / 8, msaaLevel / 16, msaaLevel / 32, 0};
-	const std::array zbBits = {24, 32, 16};
-
-	const char* wpfName = "";
-
-	const char* frmts[2] = {
-		"[GR::%s] error \"%s\" using %dx anti-aliasing and %d-bit depth-buffer for main window",
-		"[GR::%s] using %dx anti-aliasing and %d-bit depth-buffer (PF=\"%s\") for main window",
-	};
-
-	bool borderless_ = configHandler->GetBool("WindowBorderless");
-	bool fullScreen_ = configHandler->GetBool("Fullscreen");
-	int winPosX_ = configHandler->GetInt("WindowPosX");
-	int winPosY_ = configHandler->GetInt("WindowPosY");
-	int2 newRes = GetCfgWinRes();
-
-	// note:
-	//   passing the minimized-flag is useless (state is not saved if minimized)
-	//   and has no effect anyway, setting a minimum size for a window overrides
-	//   it while disabling the SetWindowMinimumSize call still results in a 1x1
-	//   window on the desktop
-	//
-	//   SDL_WINDOW_FULLSCREEN, for "real" fullscreen with a videomode change;
-	//   SDL_WINDOW_FULLSCREEN_DESKTOP for "fake" fullscreen that takes the size of the desktop;
-	//   and 0 for windowed mode.
-
-	uint32_t sdlFlags  = (SDL_WINDOW_OPENGL | SDL_WINDOW_RESIZABLE);
-	         sdlFlags |= (borderless_ ? SDL_WINDOW_FULLSCREEN_DESKTOP : SDL_WINDOW_FULLSCREEN) * fullScreen_;
-	         sdlFlags |= (SDL_WINDOW_BORDERLESS * borderless_);
-
-	for (size_t i = 0; i < (aaLvls.size()) && (newWindow == nullptr); i++) {
-		if (i > 0 && aaLvls[i] == aaLvls[i - 1])
-			break;
-
-		SDL_GL_SetAttribute(SDL_GL_MULTISAMPLEBUFFERS, aaLvls[i] > 0);
-		SDL_GL_SetAttribute(SDL_GL_MULTISAMPLESAMPLES, aaLvls[i]    );
-
-		for (size_t j = 0; j < (zbBits.size()) && (newWindow == nullptr); j++) {
-			SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, zbBits[j]);
-
-			if ((newWindow = SDL_CreateWindow(title, winPosX_, winPosY_, newRes.x, newRes.y, sdlFlags)) == nullptr) {
-				LOG_L(L_WARNING, frmts[0], __func__, SDL_GetError(), aaLvls[i], zbBits[j]);
-				continue;
-			}
-
-			LOG(frmts[1], __func__, aaLvls[i], zbBits[j], wpfName = SDL_GetPixelFormatName(SDL_GetWindowPixelFormat(newWindow)));
-		}
-	}
-
-	if (newWindow == nullptr) {
-		auto buf = fmt::sprintf("[GR::%s] could not create SDL-window\n", __func__);
-		handleerror(nullptr, buf.c_str(), "ERROR", MBF_OK | MBF_EXCL);
-		return nullptr;
-	}
-
-	UpdateWindowBorders(newWindow);
-
-	return newWindow;
-}
-
-SDL_GLContext CGlobalRendering::CreateGLContext(const int2& minCtx)
-{
-	SDL_GLContext newContext = nullptr;
-
-	constexpr int2 glCtxs[] = {{2, 0}, {2, 1},  {3, 0}, {3, 1}, {3, 2}, {3, 3},  {4, 0}, {4, 1}, {4, 2}, {4, 3}, {4, 4}, {4, 5}, {4, 6}};
-	          int2 cmpCtx;
-
-	if (std::find(&glCtxs[0], &glCtxs[0] + (sizeof(glCtxs) / sizeof(int2)), minCtx) == (&glCtxs[0] + (sizeof(glCtxs) / sizeof(int2)))) {
-		handleerror(nullptr, "illegal OpenGL context-version specified, aborting", "ERROR", MBF_OK | MBF_EXCL);
-		return nullptr;
-	}
-
-	if ((newContext = SDL_GL_CreateContext(sdlWindow)) != nullptr)
-		return newContext;
-
-	const char* frmts[] = {"[GR::%s] error (\"%s\") creating main GL%d.%d %s-context", "[GR::%s] created main GL%d.%d %s-context"};
-	const char* profs[] = {"compatibility", "core"};
-
-	char buf[1024] = {0};
-	SNPRINTF(buf, sizeof(buf), frmts[false], __func__, SDL_GetError(), minCtx.x, minCtx.y, profs[forceCoreContext]);
-
-	for (const int2 tmpCtx: glCtxs) {
-		SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, tmpCtx.x);
-		SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, tmpCtx.y);
-
-		for (uint32_t mask: {SDL_GL_CONTEXT_PROFILE_CORE, SDL_GL_CONTEXT_PROFILE_COMPATIBILITY}) {
-			SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, mask);
-
-			if ((newContext = SDL_GL_CreateContext(sdlWindow)) == nullptr) {
-				LOG_L(L_WARNING, frmts[false], __func__, SDL_GetError(), tmpCtx.x, tmpCtx.y, profs[mask == SDL_GL_CONTEXT_PROFILE_CORE]);
-			} else {
-				// save the lowest successfully created fallback compatibility-context
-				if (mask == SDL_GL_CONTEXT_PROFILE_COMPATIBILITY && cmpCtx.x == 0 && tmpCtx.x >= minCtx.x)
-					cmpCtx = tmpCtx;
-
-				LOG_L(L_WARNING, frmts[true], __func__, tmpCtx.x, tmpCtx.y, profs[mask == SDL_GL_CONTEXT_PROFILE_CORE]);
-			}
-
-			// accepts nullptr's
-			SDL_GL_DeleteContext(newContext);
-		}
-	}
-
-	if (cmpCtx.x == 0) {
-		handleerror(nullptr, buf, "ERROR", MBF_OK | MBF_EXCL);
-		return nullptr;
-	}
-
-	SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, cmpCtx.x);
-	SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, cmpCtx.y);
-	SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_COMPATIBILITY);
-
-	// should never fail at this point
-	return (newContext = SDL_GL_CreateContext(sdlWindow));
-}
-
-bool CGlobalRendering::CreateWindowAndContext(const char* title)
-{
-	if (SDL_Init(SDL_INIT_VIDEO) == -1) {
-		LOG_L(L_FATAL, "[GR::%s] error \"%s\" initializing SDL", __func__, SDL_GetError());
-		return false;
-	}
-
-	if (!CheckAvailableVideoModes()) {
-		handleerror(nullptr, "desktop color-depth should be at least 24 bits per pixel, aborting", "ERROR", MBF_OK | MBF_EXCL);
-		return false;
-	}
-
-	// should be set to "3.0" (non-core Mesa is stuck there), see below
-	const char* mesaGL = getenv("MESA_GL_VERSION_OVERRIDE");
-	const char* softGL = getenv("LIBGL_ALWAYS_SOFTWARE");
-
-	// get wanted resolution and context-version
-	const int2 minCtx = (mesaGL != nullptr && std::strlen(mesaGL) >= 3)?
-		int2{                  std::max(mesaGL[0] - '0', 3),                   std::max(mesaGL[2] - '0', 0)}:
-		int2{configHandler->GetInt("GLContextMajorVersion"), configHandler->GetInt("GLContextMinorVersion")};
-
-	// start with the standard (R8G8B8A8 + 24-bit depth + 8-bit stencil + DB) format
-	SDL_GL_SetAttribute(SDL_GL_RED_SIZE,   8);
-	SDL_GL_SetAttribute(SDL_GL_GREEN_SIZE, 8);
-	SDL_GL_SetAttribute(SDL_GL_BLUE_SIZE,  8);
-	SDL_GL_SetAttribute(SDL_GL_ALPHA_SIZE, 8);
-	SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE,  24);
-	SDL_GL_SetAttribute(SDL_GL_STENCIL_SIZE, 8);
-	SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER, 1);
-
-	// create GL debug-context if wanted (more verbose GL messages, but runs slower)
-	// note:
-	//   requesting a core profile explicitly is needed to get versions later than
-	//   3.0/1.30 for Mesa, other drivers return their *maximum* supported context
-	//   in compat and do not make 3.0 itself available in core (though this still
-	//   suffices for most of Spring)
-	SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, forceCoreContext? SDL_GL_CONTEXT_PROFILE_CORE: SDL_GL_CONTEXT_PROFILE_COMPATIBILITY);
-	SDL_GL_SetAttribute(SDL_GL_CONTEXT_FLAGS, SDL_GL_CONTEXT_DEBUG_FLAG);
-
-	SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, minCtx.x);
-	SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, minCtx.y);
-
-
-	if (msaaLevel > 0) {
-		if (softGL != nullptr)
-			LOG_L(L_WARNING, "MSAALevel > 0 and LIBGL_ALWAYS_SOFTWARE set, this will very likely crash!");
-
-		// has to be even
-		if (msaaLevel % 2 == 1)
-			++msaaLevel;
-	}
-
-	if ((sdlWindow = CreateSDLWindow(title)) == nullptr)
-		return false;
-
-	if (configHandler->GetInt("MinimizeOnFocusLoss") == 0)
-		SDL_SetHint(SDL_HINT_VIDEO_MINIMIZE_ON_FOCUS_LOSS, "0");
-
-	SetWindowAttributes(sdlWindow);
-
-#if !defined(HEADLESS)
-	// disable desktop compositing to fix tearing
-	// (happens at 300fps, neither fullscreen nor vsync fixes it, so disable compositing)
-	// On Windows Aero often uses vsync, and so when Spring runs windowed it will run with
-	// vsync too, resulting in bad performance.
-	if (configHandler->GetBool("BlockCompositing"))
-		WindowManagerHelper::BlockCompositing(sdlWindow);
-#endif
-
-	if ((glContext = CreateGLContext(minCtx)) == nullptr)
-		return false;
-
-	gladLoadGL();
-	GLX::Load(sdlWindow);
-
-	if (!CheckGLContextVersion(minCtx)) {
-		int ctxProfile = 0;
-		SDL_GL_GetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, &ctxProfile);
-
-		const std::string errStr = fmt::format("current OpenGL version {}.{}(core={}) is less than required {}.{}(core={}), aborting",
-			globalRenderingInfo.glContextVersion.x, globalRenderingInfo.glContextVersion.y, globalRenderingInfo.glContextIsCore,
-			minCtx.x, minCtx.y, (ctxProfile == SDL_GL_CONTEXT_PROFILE_CORE)
-		);
-
-		handleerror(nullptr, errStr.c_str(), "ERROR", MBF_OK | MBF_EXCL);
-		return false;
-	}
-
-	MakeCurrentContext(false);
-	SDL_DisableScreenSaver();
-	return true;
-}
-
-
 void CGlobalRendering::MakeCurrentContext(bool clear) const {
 	SDL_GL_MakeCurrent(sdlWindow, clear ? nullptr : glContext);
-}
-
-
-void CGlobalRendering::DestroyWindowAndContext() {
-	if (!sdlWindow)
-		return;
-
-	WindowManagerHelper::SetIconSurface(sdlWindow, nullptr);
-	SetWindowInputGrabbing(false);
-
-	SDL_GL_MakeCurrent(sdlWindow, nullptr);
-	SDL_DestroyWindow(sdlWindow);
-
-	#if !defined(HEADLESS)
-	if (glContext)
-		SDL_GL_DeleteContext(glContext);
-	#endif
-
-	sdlWindow = nullptr;
-	glContext = nullptr;
-
-	GLX::Unload();
-}
-
-void CGlobalRendering::KillSDL() const {
-	#if !defined(HEADLESS)
-	SDL_QuitSubSystem(SDL_INIT_VIDEO);
-	#endif
-
-	SDL_EnableScreenSaver();
-	SDL_Quit();
 }
 
 void CGlobalRendering::PostInit() {
